requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Deprecated==1.2.10
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
PyGithub==1.52
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
python-dotenv==0.14.0
regex==2020.7.14
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
zipp==3.1.0
